Setup the method for getting user from db
A problem with the implementation so far is that we would have to make
multiple calls to the slack and google api for the user information we
need. Even that will lead to the problem of getting every single event
from the calendar and trying to match that with the user info.

 This commit will mean that the the relevant user info will stay in the
db and the created hash for ease/efficiency of reach.